# [10.x] Improve `collect()->push()` performance