@michaelnabil230 2024-01-01
Hi @GrahamCampbell, Ummm... I don't have any solution now, but do you think that this style of writing is widespread in use, and if it is widespread, do you have any solutions to this problem?